+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.3.8] - 2020-03-04
+### Added
+- Australia region support
+
 ### Fixed
 - The similarity report uses the role that is defined in the course context